requirements.txt
addict==2.4.0
anyconfig==0.14.0
apted==1.0.3
beautifulsoup4==4.12.3
Cython==3.0.9
docx==0.2.4
editdistance==0.8.1
fastapi==0.111.0
fasttext==0.9.2
imgaug==0.4.0
lanms==1.0.2
lmdb==1.4.1
lxml==4.9.3
matplotlib==3.7.3
numpy==1.24.4
onnxruntime==1.18.0
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
openpyxl==3.1.3
paddleclas==2.5.2
paddlenlp==2.8.0
paddlepaddle==2.6.1
paddleslim==2.6.0
pdf2docx==0.5.8
Pillow==10.3.0
Polygon==1.2.0
premailer==3.10.0
pyclipper==1.3.0.post5
pydantic==1.10.15
pytest==8.2.2
python-dotenv==1.0.1
PyYAML==6.0.1
PyYAML==6.0.1
QtPy==2.4.1
rapidfuzz==3.9.3
Requests==2.32.3
scipy==1.13.1
seqeval==1.2.2
setuptools==68.2.2
Shapely==2.0.4
six==1.16.0
skimage==0.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.65.0
uvicorn==0.30.1
visualdl==2.5.3
wandb==0.17.1
